+++ client/src/components/LoginModal/index.js
@@ -2,16 +2,16 @@ import React from "react";
 // import { Link } from "react-router-dom";
 import { Typography, Box } from "@mui/material";
 import Grid from "@mui/material/Unstable_Grid2";
-import SignupContainer from "./SignupContainer";
+import LoginContainer from "./LoginContainer";
 
 const SignupModal = () => {
 	return (
 		<Box elevation={2} sx={{ flexGrow: 1 }}>
 			<Typography variant="h1" component="h1">
-				Signup
+				Login
 			</Typography>
 			<Grid container>
-				<SignupContainer />
+				<LoginContainer />
 			</Grid>
 		</Box>
 	);
