--- client/src/components/Nav/header.css
@@ -1,9 +1,0 @@
-.header {
-  display: flex;
-  justify-content: space-between;
-  background-color: red;
-  padding: 18px 10px;
-  color: #fff;
-  font-size: 24px;
-  min-height: 20px;
-}
