+++ client/src/pages/Login.js
@@ -1,0 +1,6 @@
+import LoginModal from "../components/LoginModal";
+const Signup = () => {
+	return <LoginModal />;
+};
+
+export default Signup;
